# src/main/java/com/example/store/dto/ProductDTO.java
package com.example.store.dto;

import lombok.Data;

import java.util.List;

/** Product dto. */
@Data
public class ProductDTO {
    private Long id;
    private String description;
    private List<Long> orderIds;
}